+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog for aim.models
 
 - Added default to IS3Bucket.policy
 - Added Route53 to schema and model
+- Removed unused yaml config from aimdemo under fixtures.
 
 
 1.0.1 (2019-06-19)